Copy rather than symlink user-modifiable files
This commit makes the setup script copy the following files into the
project repository, rather than symbolically linking them:

  * Makefile-project-defines
  * Makefile-project-targets
  * Makefile-project-testing

This is because project owners are expected to modify these files in the
project repository. These changes should not be written to the
submodule, because the submodule would then become dirty and impossible
to update.